move failover.disabled_reasons to own module
