+++ src/middlewared/middlewared/plugins/failover_/disabled_reasons.py
@@ -1,0 +1,80 @@
+from middlewared.schema import accepts, returns, List, Str
+from middlewared.service import ConfigService, throttle, pass_app, no_auth_required
+from middlewared.plugins.failover_.utils import throttle_condition
+
+
+class FailoverService(ConfigService):
+
+    LAST_DISABLEDREASONS = None
+
+    @no_auth_required
+    @throttle(seconds=2, condition=throttle_condition)
+    @accepts()
+    @returns(List('reasons', items=[Str('reason')]))
+    @pass_app()
+    def disabled_reasons(self, app):
+        """
+        Returns a list of reasons why failover is not enabled/functional.
+
+        NO_VOLUME - There are no pools configured.
+        NO_VIP - There are no interfaces configured with Virtual IP.
+        NO_SYSTEM_READY - Other storage controller has not finished booting.
+        NO_PONG - Other storage controller is not communicable.
+        NO_FAILOVER - Failover is administratively disabled.
+        NO_LICENSE - Other storage controller has no license.
+        DISAGREE_CARP - Nodes CARP states do not agree.
+        MISMATCH_DISKS - The storage controllers do not have the same quantity of disks.
+        NO_CRITICAL_INTERFACES - No network interfaces are marked critical for failover.
+        """
+        reasons = set(self._disabled_reasons(app))
+        if reasons != self.LAST_DISABLEDREASONS:
+            self.LAST_DISABLEDREASONS = reasons
+            self.middleware.send_event(
+                'failover.disabled_reasons', 'CHANGED',
+                fields={'disabled_reasons': list(reasons)}
+            )
+        return list(reasons)
+
+    def _disabled_reasons(self, app):
+        reasons = []
+        if not self.middleware.call_sync('pool.query'):
+            reasons.append('NO_VOLUME')
+        if not any(filter(
+            lambda x: x.get('failover_virtual_aliases'), self.middleware.call_sync('interface.query'))
+        ):
+            reasons.append('NO_VIP')
+        try:
+            assert self.middleware.call_sync('failover.remote_connected') is True
+
+            # if the remote node panic's (this happens on failover event if we cant export the
+            # zpool in 4 seconds on freeBSD systems (linux reboots silently by design)
+            # then the p2p interface stays "UP" and the websocket remains open.
+            # At this point, we have to wait for the TCP timeout (60 seconds default).
+            # This means the assert line up above will return `True`.
+            # However, any `call_remote` method will hang because the websocket is still
+            # open but hasn't closed due to the default TCP timeout window. This can be painful
+            # on failover events because it delays the process of restarting services in a timely
+            # manner. To work around this, we place a `timeout` of 5 seconds on the system.ready
+            # call. This essentially bypasses the TCP timeout window.
+            if not self.middleware.call_sync('failover.call_remote', 'system.ready', [], {'timeout': 5}):
+                reasons.append('NO_SYSTEM_READY')
+
+            if not self.middleware.call_sync('failover.call_remote', 'failover.licensed'):
+                reasons.append('NO_LICENSE')
+
+            local = self.middleware.call_sync('failover.vip.get_states')
+            remote = self.middleware.call_sync('failover.call_remote', 'failover.vip.get_states')
+            if self.middleware.call_sync('failover.vip.check_states', local, remote):
+                reasons.append('DISAGREE_CARP')
+
+            mismatch_disks = self.middleware.call_sync('failover.mismatch_disks')
+            if mismatch_disks['missing_local'] or mismatch_disks['missing_remote']:
+                reasons.append('MISMATCH_DISKS')
+
+            if not self.middleware.call_sync('datastore.query', 'network.interfaces', [['int_critical', '=', True]]):
+                reasons.append('NO_CRITICAL_INTERFACES')
+        except Exception:
+            reasons.append('NO_PONG')
+        if self.middleware.call_sync('failover.config')['disabled']:
+            reasons.append('NO_FAILOVER')
+        return reasons
