+++ src/middlewared/middlewared/plugins/failover_/utils.py
@@ -1,0 +1,5 @@
+def throttle_condition(middleware, app, *args, **kwargs):
+    # app is None means internal middleware call
+    if app is None or (app and app.authenticated):
+        return True, 'AUTHENTICATED'
+    return False, None
